Modulated the home page (Separated Layout Components)
